Remove API Gateway command; add auth ingresses and plugins
This commit removes the command to delete the API Gateway and adds a command to delete all ingresses in the `eshop-microservices` namespace. New ingress resources for `auth-api-login` and `auth-api-register` are introduced in `kong-ingress.yaml`, each with specific annotations and rules. Additionally, a new Kong plugin for transforming request paths for the login and register APIs replaces the previous public whitelist plugin in `kong-plugins.yaml`.